requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.17.0
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==24.3.0
bidict==0.23.1
blurhash==1.1.4
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
Mastodon.py==1.8.1
multidict==6.1.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pyenchant==3.2.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-magic==0.4.27
python-socketio==5.12.1
requests==2.32.3
schedule==1.2.2
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.3.0
wsproto==1.2.0
yarl==1.18.3
